requirements.txt
Flask==1.0.2
pillow==5.4.0
requests==2.21.0
